Add hex colour codes to README
I needed tooltips which had the same colours as the markers, so for the
CSS I needed to know the colours inside and outside of the markers.

It would have been helpful for me to have had the colours listed so I
could use them. I have now determined these colour and put them in the
README in the hope that they may be useful to someone else with similar
requirements.

I determined these colours by using a colour picker on the icons, so
they may not be 100% accurate, but they're good enough for my
application.